formularios de: estado cita y especialidades control de errores revisar para registrar una cita

+++ src/main/java/com/tii/springboot/app/models/dao/IEstadoCitaDao.java
@@ -10,6 +10,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
  *
  * @author Pe-Hs
  */
-public interface IEstadoCitaDao extends PagingAndSortingRepository<EstadoCita, Integer>{
+public interface IEstadoCitaDao extends PagingAndSortingRepository<EstadoCita, Long>{
     
 }
